+++ R/race-wrapper.R
@@ -208,7 +208,7 @@ target.evaluator.default <- function(experiment, num.configurations, all.conf.id
 
   cwd <- setwd (scenario$execDir)
   args <- c(configuration.id, instance.id, seed, instance, num.configurations, all.conf.id)
-  output <- runcommand(targetEvaluator, args, configuration.id, debugLevel, use_std = use_std)
+  output <- runcommand(targetEvaluator, args, configuration.id, debugLevel, use_std = use_std, timeout = scenario$targetRunnerTimeout)
   setwd (cwd)
 
   cost <- time <- NULL
@@ -438,7 +438,7 @@ run_target_runner <- function(experiment, scenario)
     error <- "targetRunnerLauncher"
   }
   if (!use_std) {file.check(targetRunner, executable = TRUE, text = error)}
-  output <- runcommand(targetRunner, args, configuration.id, debugLevel, use_std = use_std)
+  output <- runcommand(targetRunner, args, configuration.id, debugLevel, use_std = use_std, timeout = scenario$targetRunnerTimeout)
   list(cmd=targetRunner, output=output, args=args)
 }