+++ NEWS.md
@@ -57,6 +57,12 @@
    `repairConfiguration` function (see `defaultScenario()`) only needs two
    arguments: `configuration` and `parameters`. See examples in the user-guide.
 
+ * `readScenario()` (and command-line irace) do not require a `scenario.txt` file. (Contributed by @DE0CH)
+ 
+ * Command-line options in joined form, given as "--log-file= --check", without any argument after the '=' will be interpreted as an empty argument,  equivalent to using `logFile=""` in `scenario.txt`.
+
+ * New scenario option `targetRunnerTimeout`: Timeout in seconds of any `targetRunner` call (only applies to `target-runner` executables not to R functions).
+
  * New options `trainInstancesText`, `testInstancesText`, `parameterText`, and
    `configurationsText` which accepts the content that is normally in
    `trainInstancesFile`, `testInstancesFile`, `parameterText`, and 
@@ -64,7 +70,11 @@
 
 ## Fixes
 
- * Fix (#44): irace now will give an error if the domain of real-valued (r)
+ * Fix #55: Configurations provided may use `<NA>` in addition to `NA` to denote
+   the missing value of a disabled parameter.
+                                      (Manuel López-Ibáñez, reported by @TheIronBorn)
+ 
+ * Fix #44: irace now will give an error if the domain of real-valued (r)
    parameters would change depending on the value of `'digits'`. The solution
    is to increase the value of `'digits'` or adjust the domain. 
                                        (Manuel López-Ibáñez, reported by @mb706)